add test for adding suppressed exceptions for haltOnParseError
